Updated query namespaces to plural
