+++ backend/src/Equinor.ProjectExecutionPortal.Application/Queries/OnboardedApps/OnboardedAppDto.cs
@@ -1,7 +1,7 @@
 ﻿using Equinor.ProjectExecutionPortal.Application.Infrastructure.Mappings;
-using Equinor.ProjectExecutionPortal.Application.Queries.AppGroup;
+using Equinor.ProjectExecutionPortal.Application.Queries.AppGroups;
 
-namespace Equinor.ProjectExecutionPortal.Application.Queries.OnboardedApp;
+namespace Equinor.ProjectExecutionPortal.Application.Queries.OnboardedApps;
 
 public class OnboardedAppDto : IMapFrom<Domain.Entities.OnboardedApp>
 {
