+++ backend/src/Equinor.ProjectExecutionPortal.Application/Queries/OnboardedContexts/OnboardedContextDto.cs
@@ -1,6 +1,8 @@
 ﻿using Equinor.ProjectExecutionPortal.Application.Infrastructure.Mappings;
 using Equinor.ProjectExecutionPortal.Domain.Entities;
 
+namespace Equinor.ProjectExecutionPortal.Application.Queries.OnboardedContexts;
+
 public class OnboardedContextDto : IMapFrom<OnboardedContext>
 {
     public Guid Id { get; set; }
