+++ backend/src/Equinor.ProjectExecutionPortal.Application/Queries/WorkSurfaces/GetWorkSurfaceAppGroupsWithGlobalApps/GetWorkSurfaceAppGroupsWithGlobalAppsQuery.cs
@@ -8,20 +8,18 @@ using Equinor.ProjectExecutionPortal.Infrastructure;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
-namespace Equinor.ProjectExecutionPortal.Application.Queries.WorkSurface.GetWorkSurfaceAppGroupsWithContextAndGlobalApps;
+namespace Equinor.ProjectExecutionPortal.Application.Queries.WorkSurfaces.GetWorkSurfaceAppGroupsWithGlobalApps;
 
-public class GetWorkSurfaceAppGroupsWithContextAndGlobalAppsQuery : QueryBase<IList<WorkSurfaceAppGroupWithAppsDto>>
+public class GetWorkSurfaceAppGroupsWithGlobalAppsQuery : QueryBase<IList<WorkSurfaceAppGroupWithAppsDto>>
 {
-    public GetWorkSurfaceAppGroupsWithContextAndGlobalAppsQuery(Guid workSurfaceId, string? contextExternalId)
+    public GetWorkSurfaceAppGroupsWithGlobalAppsQuery(Guid workSurfaceId)
     {
         WorkSurfaceId = workSurfaceId;
-        ContextExternalId = contextExternalId;
     }
 
     public Guid WorkSurfaceId { get; }
-    public string? ContextExternalId { get; }
 
-    public class Handler : IRequestHandler<GetWorkSurfaceAppGroupsWithContextAndGlobalAppsQuery, IList<WorkSurfaceAppGroupWithAppsDto>>
+    public class Handler : IRequestHandler<GetWorkSurfaceAppGroupsWithGlobalAppsQuery, IList<WorkSurfaceAppGroupWithAppsDto>>
     {
         private readonly IReadWriteContext _readWriteContext;
         private readonly IAppService _appService;
@@ -36,12 +34,12 @@ public class GetWorkSurfaceAppGroupsWithContextAndGlobalAppsQuery : QueryBase<IL
             _mapper = mapper;
         }
 
-        public async Task<IList<WorkSurfaceAppGroupWithAppsDto>> Handle(GetWorkSurfaceAppGroupsWithContextAndGlobalAppsQuery request, CancellationToken cancellationToken)
+        public async Task<IList<WorkSurfaceAppGroupWithAppsDto>> Handle(GetWorkSurfaceAppGroupsWithGlobalAppsQuery request, CancellationToken cancellationToken)
         {
             var workSurface = await _readWriteContext.Set<Domain.Entities.WorkSurface>()
                 .AsNoTracking()
-                .Include(workSurface => workSurface.Apps.Where(app => app.OnboardedContext == null || app.OnboardedContext.ExternalId == request.ContextExternalId))
-                .ThenInclude(app => app.OnboardedApp)
+                .Include(workSurface => workSurface.Apps.Where(app => app.OnboardedContextId == null))
+                .ThenInclude(workSurfaceApp => workSurfaceApp.OnboardedApp)
                 .ThenInclude(onboardedApp => onboardedApp.AppGroup)
                 .FirstOrDefaultAsync(x => x.Id == request.WorkSurfaceId, cancellationToken);
 
@@ -51,11 +49,11 @@ public class GetWorkSurfaceAppGroupsWithContextAndGlobalAppsQuery : QueryBase<IL
             }
 
             var workSurfaceDto = _mapper.Map<Domain.Entities.WorkSurface, WorkSurfaceDto>(workSurface);
-            
+
             await _appService.EnrichAppsWithFusionAppData(workSurfaceDto.Apps.Select(workSurfaceAppDto => workSurfaceAppDto.OnboardedApp).ToList(), cancellationToken);
 
             var appGroupsWithApps = _workSurfaceService.MapWorkSurfaceToAppGroups(workSurfaceDto);
-            
+
             return appGroupsWithApps;
         }
     }
