+++ backend/src/Equinor.ProjectExecutionPortal.Application/Services/AppService/AppService.cs
@@ -1,5 +1,5 @@
 ﻿using Equinor.ProjectExecutionPortal.Application.Cache;
-using Equinor.ProjectExecutionPortal.Application.Queries.OnboardedApp;
+using Equinor.ProjectExecutionPortal.Application.Queries.OnboardedApps;
 using Equinor.ProjectExecutionPortal.FusionPortalApi.Apps.Models;
 
 namespace Equinor.ProjectExecutionPortal.Application.Services.AppService
