+++ backend/src/Equinor.ProjectExecutionPortal.WebApi/Controllers/PortalController.cs
@@ -1,4 +1,4 @@
-﻿using Equinor.ProjectExecutionPortal.Application.Queries.Portal.GetPortalWithApps;
+﻿using Equinor.ProjectExecutionPortal.Application.Queries.Portals.GetPortalWithApps;
 using Equinor.ProjectExecutionPortal.Application.Services.AppService;
 using Equinor.ProjectExecutionPortal.FusionPortalApi.Apps.Models;
 using Equinor.ProjectExecutionPortal.WebApi.ViewModels.Portal;
