+++ backend/src/Equinor.ProjectExecutionPortal.WebApi/ViewModels/OnboardedApp/ApiOnboardedApp.cs
@@ -1,4 +1,4 @@
-﻿using Equinor.ProjectExecutionPortal.Application.Queries.OnboardedApp;
+﻿using Equinor.ProjectExecutionPortal.Application.Queries.OnboardedApps;
 using Equinor.ProjectExecutionPortal.WebApi.ViewModels.AppGroup;
 
 namespace Equinor.ProjectExecutionPortal.WebApi.ViewModels.OnboardedApp
